fix: robot appearing on other screens and taking over space
